Fixing Bounce pads, exceptional multipliers still an issue
When bouncing from a bounce pad, if the player collides with another pad during the jump, it will multiply and fly them at a high velocity in the direction they are traveling. Often smacks them against the ceiling of the level. We'll need a way to reset velocity before enacting the boucepad multiplier